Remove AndroidTouchHandler interface
